requirements.txt
anthropic>=0.52.0
sentence-transformers>=3.0.0
numpy>=1.26.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
